% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparables_days_means.R
\name{comparables.days.means}
\alias{comparables.days.means}
\title{Compara promedio de dias especiales}
\usage{
comparables.days.means(
  df,
  fields.date,
  valor,
  fecha,
  agrupador,
  festividad,
  mov.days = NULL,
  graf.label = FALSE
)
}
\arguments{
\item{df}{nombre del data.frame a analizar}

\item{fields.date}{el nombre del campo de fecha.}

\item{valor}{El valor con el cuál calcular los promedios.}

\item{fecha}{vector con las fechas comerciales, es decir c('fecha1','fecha2').}

\item{agrupador}{El nivel de agrupamiento (por ej: Merchant, Pais, Región).}

\item{festividad}{el nombre de la fecha comercial.}

\item{mov.days}{las fechas en que sucedió el evento anterior como vector, es decir c('fecha1','fecha2')}

\item{graf.label}{booleano para indicar si queremos que agregue el valor encima de la barra poner TRUE}
}
\value{
cuadro que contiene el promedio de la fecha en estudio y los valores absolutos y porcentuales
de la comparación con semana previa (pw), mes previo (pm), año previo (py) y si no es fija (pmov)

grafico de barras con las comparaciones temporales
}
\description{
La función comparables_days_means calcula el promedio del período de
la fecha comercial y de la semana, mes y año previo. Ofrece también
la posibilidad de analizar el período anterior en caso que la fecha no
siempre coincida en la misma altura del año.
}
