% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_calendar.R
\name{heatmap.calendar}
\alias{heatmap.calendar}
\title{Crea un mapa de calor sobre un calendario}
\usage{
heatmap.calendar(
  df,
  fini,
  ffin,
  fields.date,
  agrupador,
  filtro = NULL,
  valor,
  titulo,
  exportar = FALSE
)
}
\arguments{
\item{df}{nombre del data.frame a analizar.}

\item{fini}{fecha de inicio de la secuencia en formato (yyyy-mm-dd).}

\item{ffin}{fecha de fin de la secuencia en formato (yyyy-mm-dd).}

\item{fields.date}{nombre del campo de fecha.}

\item{agrupador}{El nivel de agrupamiento (por ej: Merchant, Pais, Región).}

\item{filtro}{NULL. Si desea aplicar un filtro al campo agrupador indicar el valor.}

\item{valor}{nombre del campo con el valor a analizar graficamente.}

\item{titulo}{string con el titulo que desee para el gráfico.}

\item{exportar}{FALSE. Si desea exportar el archivo en formato .png cambie a TRUE.}
}
\value{
grafico de mapa de calor sobre un calendario.
}
\description{
La función heatmap_calendar genera un mapa de calor impreso sobre un calendario.
De esa forma obtenemos un objeto gráfico que representa con un color más intenso
aquéllos días con valores más altos y un color más débil para los días con valores
más bajos.
}
