% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_report.R
\name{print_report}
\alias{print_report}
\title{Guarda el reporte en un archivo excel}
\usage{
print_report(ls, festividad, name, sheetname, append = FALSE)
}
\arguments{
\item{ls}{objeto lista generado por la funci贸n \emph{comparables_days_means}.}

\item{festividad}{nombre de la fecha especial analizada.}

\item{name}{nombre del archivo.}

\item{sheetname}{nombre de la sheet. Importante cuando decidimos anexar
nuestro reporte a un archivo excel existente.}

\item{append}{FALSE. Si queremos anexarlo a un archivo existente cambiar a TRUE.}
}
\value{
archivo excel en la misma ubicaci贸n del proyecto.
}
\description{
La funci贸n print_report permite guardar los productos de la funci贸n
\strong{comparables_days_means} en un archivo excel. Podemos guardarlo en un
nuevo archivo o anexarlo a uno existente.
}
