% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventas.R
\docType{data}
\name{ventas}
\alias{ventas}
\title{Ventas}
\format{
Data frame con 4 columnas y 1.580 filas.
\describe{
\item{fechas}{formato (yyyy-mm-dd)}
\item{merchant}{nombre del local}
\item{country}{pais dónde se encuentra el local}
\item{cantidad}{cantidad de bienes vendidos}
}
}
\usage{
ventas
}
\description{
Contiene las ventas de dos locales en Argentina y Uruguay entre el 1 de junio de 2021 y el 30 de junio de 2022.
}
\examples{
  ventas
}
\keyword{datasets}
